+++ src/components/shared/applied-filters.tsx
